+++ dwds/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.7.5-dev
 
+- The injected client's connection is now based off the request URI.
 - Fix an issue where resuming while paused at the start would cause an error.
 - Expose the `ChromeDebugException` class for error handling purposes.